Parent frame in select menu set
The frame that contains all of the optimization inner frames to be displayed has been correctly configured aesthetically. However, although the scrollregion seems to be correct vertically, one cannot scroll unless the cursor is over the scrollbar, so that needs to be corrected. Also, some corrections regarding the files taken into account to set the current database file name were done to the get_last_xlsx_file method. Added comments to data.py to bear in mind when executing the GUI as an attribute of Main