Implemented KeyRealese-Return
For the info button in the help frame, the KeyRelease-Return event was added. This allows the user to enter the info frame using only the tab to move around the buttons and the enter key to hide or open the frame. This will be an extra feature to be added to the final implementation. For now, the focus will turn to converting the GUI to multiple classes.