Defined BottomFrame class
In the module bottomFrame the new class of main's bottom frame can be found. It was decided to divide them as such to have better modularity with that menu and to also set the optimization button frame in another module to have better organized code. That way, MainMenu will contain objects of the two framesand the title, while having direct interaction with GUI. The  functionality of the frame inside the menu is left to be tested however, and can't be done until the optimization button frame has been also defined. Therefore, that's the next step.