Finished MainMenu
The main menu works correctly, but it is an object of a class which contains two objects of two other classes: OptionFrame and BottomFrame. All the buttons were correctly binded, even to the -enter- key. Left to check, however, if it works correctly in Windows, especially when using -enter- and the styles when entering, leaving, clickind and releasing the buttons. On the mean time, I will continue with the select menu implementation.