Created OptionsButton class
A new class that inherits from tk.Button was created to initialize the create, select and exit buttons of the main menu. They work properly in this menu. However, the next step must be to redifine the display method so that the create_optimization_button in SelectMenu can be an instance of this class.