PopUpFrame and BottomButton implementation
The release method of PopUpFrame's objects was succesfully tested and a few bugs with the scrollbar were corrected. In addition, a few lines were also added in other files to implement binding methods and to reset the visible area of a scrollable widget. Now, the development will be focused on the create optimzation menu and creating a BackButton class, which will probably be needed in the class of the create optimization menu.