BackButton class and inheritance added
BackButton is a new class that generates a back button from an image. It was made so that the one in the select menu were to be the same as the one in the create optimization menu. Also, the bottom and optimization frames classes now inherit from tk.Frame. Now, development will continue with the create optimization menu as planned