Started create menu
Succesfully implemented the input and their related labels, as well as an OptionMenu object to show all the functions the user can choose for the optimization. Also, a plot is now being shown in the right-hand side of the screen. It is left to be conditioned by the choice in the OptionMenu, alongside its contour levels, but it is visible!