Successfully displaying functions
In the create menu, now the dropdown menu is correctly associated with the graph displayed on the right, i.e., a choice change in the dropdown triggers a graph change. Now, the buttons below the graph are left, and they will start the interaction between all the menus and the Main object.