ViewMenu development started
The ViewMenu and the ViewButton classes have been created. The latter triggers the first one to appear, succesfully showing the particles in the first iteration of the optimization. The next step is to show different iterations as time passes and the play button is released. Also np.clip needs to be implemented in the algorithm so the particles don't appear outisde the function's plot.