Feature module packages
The graphics package (directory) was modified so it now contains three subpackage modules corresponding to the create, select and main menus. The exit menu is a single file, so it wasn't stored in a different package. Also, some requirements in the .toml file were updated to comply with poetry's own suggestions. Now, the specifications to create a venv with poetry will be written